# docs/swagger.yaml
definitions:
...
  server_model_common_request.GetById:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
  server_model_common_request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  server_model_common_response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  server_model_common_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  server_model_system.SysApi:
    properties:
      api_group:
        description: api组
        type: string
      createdAt:
        description: 创建时间
        type: string
      description:
        description: api中文描述
        type: string
      id:
        description: 主键ID
        type: integer
      method:
        description: 方法:创建POST(默认)|查看GET|更新PUT|删除DELETE
        type: string
      path:
        description: api路径
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  server_model_system_request.SearchApiParams:
    properties:
      api_group:
        description: api组
        type: string
      createdAt:
        description: 创建时间
        type: string
      desc:
        description: 排序方式:升序false(默认)|降序true
        type: string
      description:
        description: api中文描述
        type: string
      id:
        description: 主键ID
        type: integer
      is_cropper:
        description: 是否为截图 1, 或 2
        type: string
      keyword:
        description: 关键字
        type: string
      method:
        description: 方法:创建POST(默认)|查看GET|更新PUT|删除DELETE
        type: string
      orderKey:
        description: 排序
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      path:
        description: api路径
        type: string
      proportion:
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/DeleteApi/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除api
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "400":
          description: 参数错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除api
      tags:
      - SysApi
  /api/createApi:
    post:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/server_model_system.SysApi'
      produces:
      - application/json
      responses:
        "200":
          description: 创建基础api
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "400":
          description: 参数错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建基础api
      tags:
      - SysApi
  /api/deleteApisByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/server_model_common_request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 删除选中Api
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "400":
          description: 参数错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除选中Api
      tags:
      - SysApi
  /api/getAllApis:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取所有的Api 不分页,返回包括api列表
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/server_model_system.SysApi'
                  type: array
                msg:
                  type: string
              type: object
        "400":
          description: 参数错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取所有的Api 不分页
      tags:
      - SysApi
  /api/getApiById/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 根据id获取api
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/server_model_common_request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: 根据id获取api,返回包括api详情
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/server_model_system.SysApi'
                msg:
                  type: string
              type: object
        "400":
          description: 参数错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 根据id获取api
      tags:
      - SysApi
  /api/getApiList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页获取API列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/server_model_system_request.SearchApiParams'
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取API列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/server_model_common_response.PageResult'
                msg:
                  type: string
              type: object
        "400":
          description: 参数错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取API列表
  /api/updateApi/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/server_model_system.SysApi'
      produces:
      - application/json
      responses:
        "200":
          description: 修改基础api
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "400":
          description: 参数错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 服务器错误
          schema:
            allOf:
            - $ref: '#/definitions/server_model_common_response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 修改基础api
...
swagger: "2.0"